added color and default message for empty list
